[RoboCIn] Refactor: reorganize files
